Split game logic from GUI
The game logic is split from the GUI. This is necessary to make testing
easier (as it is hard to test GUIs). The new class Engine stores all the
Entities, and the Entities themselves store their position and
size. Every frame, the GUI requests the Engine to update the positions
of all Entities. The added and removed entities are requested by the
GUI. The GUI should create new sprites from the added entities, and the
GUI should remove sprites based on the removed entities.