Fix multiplayer saves
Previously, the save file did not contain whether this game is a
multiplayer game or not. After one presses "Continue" in the store, the
store generates a regular singleplayer PlayScene.

To fix this, it is stored in GameInformation whether this game is a
multiplayer game. The store decides based on this to create a new
PlayScene or MultiplayerPlayScene.

<bug,documentation